feat: form quick action buttons (#1140)
AB#104815

---------
