+++ src/schema/mutation/editForm.mutation.ts
@@ -1,32 +1,32 @@
+import { status, StatusEnumType, StatusType } from '@const/enumTypes';
+import { Channel, Form, ReferenceData, Resource, Version } from '@models';
+import { graphQLAuthCheck } from '@schema/shared';
+import { AppAbility } from '@security/defineUserAbility';
+import { Context } from '@server/apollo/context';
+import { logger } from '@services/logger.service';
 import {
-  GraphQLNonNull,
-  GraphQLID,
-  GraphQLString,
+  addField,
+  extractFields,
+  findDuplicateFields,
+  removeField,
+  replaceField,
+} from '@utils/form';
+import checkDefaultFields from '@utils/form/checkDefaultFields';
+import { validateGraphQLTypeName } from '@utils/validators';
+import {
   GraphQLError,
+  GraphQLID,
+  GraphQLNonNull,
+  GraphQLString,
 } from 'graphql';
 import GraphQLJSON from 'graphql-type-json';
-import { Form, Resource, Version, Channel, ReferenceData } from '@models';
-import {
-  removeField,
-  addField,
-  replaceField,
-  findDuplicateFields,
-  extractFields,
-} from '@utils/form';
-import { FormType } from '../types';
-import { validateGraphQLTypeName } from '@utils/validators';
-import mongoose from 'mongoose';
-import { AppAbility } from '@security/defineUserAbility';
-import { status, StatusEnumType, StatusType } from '@const/enumTypes';
-import isEqual from 'lodash/isEqual';
-import differenceWith from 'lodash/differenceWith';
-import unionWith from 'lodash/unionWith';
 import i18next from 'i18next';
 import { get, isArray } from 'lodash';
-import { logger } from '@services/logger.service';
-import checkDefaultFields from '@utils/form/checkDefaultFields';
-import { graphQLAuthCheck } from '@schema/shared';
-import { Context } from '@server/apollo/context';
+import differenceWith from 'lodash/differenceWith';
+import isEqual from 'lodash/isEqual';
+import unionWith from 'lodash/unionWith';
+import mongoose from 'mongoose';
+import { FormType } from '../types';
 
 /**
  * List of keys of the structure's object which we want to inherit to the children forms when they are modified on the core form
