+++ src/schema/mutation/editPage.mutation.ts
@@ -1,20 +1,22 @@
+import { contentType } from '@const/enumTypes';
+import { Button, Dashboard, Form, Page, Workflow } from '@models';
+import ActionButtonInputType from '@schema/inputs/button-action.input';
+import { graphQLAuthCheck } from '@schema/shared';
+import extendAbilityForPage from '@security/extendAbilityForPage';
+import { Context } from '@server/apollo/context';
+import { logger } from '@services/logger.service';
 import {
-  GraphQLNonNull,
-  GraphQLID,
-  GraphQLString,
-  GraphQLError,
   GraphQLBoolean,
+  GraphQLError,
+  GraphQLID,
+  GraphQLList,
+  GraphQLNonNull,
+  GraphQLString,
 } from 'graphql';
 import GraphQLJSON from 'graphql-type-json';
-import { contentType } from '@const/enumTypes';
-import { PageType } from '../types';
-import { Page, Workflow, Dashboard, Form } from '@models';
 import { cloneDeep, has, isArray, isEmpty, isNil, omit } from 'lodash';
-import extendAbilityForPage from '@security/extendAbilityForPage';
-import { logger } from '@services/logger.service';
-import { graphQLAuthCheck } from '@schema/shared';
 import { Types } from 'mongoose';
-import { Context } from '@server/apollo/context';
+import { PageType } from '../types';
 
 /** Simple form permission change type */
 type SimplePermissionChange =
@@ -39,6 +41,7 @@ type EditPageArgs = {
   permissions?: any;
   icon?: string;
   visible?: boolean;
+  buttons?: Button[];
 };
 
 /**
@@ -55,6 +58,7 @@ export default {
     icon: { type: GraphQLString },
     permissions: { type: GraphQLJSON },
     visible: { type: GraphQLBoolean },
+    buttons: { type: new GraphQLList(ActionButtonInputType) },
   },
   async resolve(parent, args: EditPageArgs, context: Context) {
     graphQLAuthCheck(context);
@@ -89,8 +93,11 @@ export default {
         ...(args.name && { name: args.name }),
         ...(args.icon && { icon: args.icon }),
         ...(has(args, 'showName') && { showName: args.showName }),
-      };
-
+      } as any;
+      // Update buttons
+      if (args.buttons) {
+        update.buttons = args.buttons;
+      }
       // Updating permissions
       const permissionsUpdate: any = {};
       if (args.permissions) {
