+++ src/schema/types/step.type.ts
@@ -1,16 +1,17 @@
-import {
-  GraphQLObjectType,
-  GraphQLID,
-  GraphQLString,
-  GraphQLBoolean,
-} from 'graphql';
-import { AccessType, WorkflowType } from '.';
+import { accessibleBy } from '@casl/mongoose';
 import { ContentEnumType } from '@const/enumTypes';
 import { Workflow } from '@models';
 import { AppAbility } from '@security/defineUserAbility';
 import extendAbilityForStep from '@security/extendAbilityForStep';
-import { accessibleBy } from '@casl/mongoose';
+import {
+  GraphQLBoolean,
+  GraphQLID,
+  GraphQLObjectType,
+  GraphQLString,
+} from 'graphql';
+import GraphQLJSON from 'graphql-type-json';
 import isNil from 'lodash/isNil';
+import { AccessType, WorkflowType } from '.';
 
 /** GraphQL Step type definition */
 export const StepType = new GraphQLObjectType({
@@ -73,6 +74,25 @@ export const StepType = new GraphQLObjectType({
       async resolve(parent, args, context) {
         const ability = await extendAbilityForStep(context.user, parent);
         return ability.can('delete', parent);
+      },
+    },
+    buttons: {
+      type: GraphQLJSON,
+      async resolve(parent, args, context) {
+        const ability = await extendAbilityForStep(context.user, parent);
+        if (ability.can('update', parent)) {
+          return parent.buttons;
+        } else {
+          return parent.buttons?.filter((button) => {
+            if (button.hasRoleRestriction) {
+              return context.user.roles?.some((role) =>
+                button.roles?.includes(role._id || '')
+              );
+            } else {
+              return true;
+            }
+          });
+        }
       },
     },
   }),
